sweet rankingの部分をコンポーネントに切り出し
